software-STM
Add methods to create queues
Synchronise the access to the inputs instruction

+++ software/code/master/Libraries/NPSC/Framework/inc/NPSC_neopixel.h
@@ -48,7 +48,7 @@
 #define RESET_SLOTS_END						(50)
 #define WS2812_LAST_SLOT					(1)
 // TODO maybe LED_BUFFER_SIZE create bugs in rings
-#define LED_BUFFER_SIZE						(RESET_SLOTS_BEGIN + LED_DATA_SIZE + WS2812_LAST_SLOT + RESET_SLOTS_END)
+#define LED_BUFFER_SIZE						(uint) (RESET_SLOTS_BEGIN + LED_DATA_SIZE + WS2812_LAST_SLOT + RESET_SLOTS_END)
 #define WS2812_0							(TIMER_PERIOD / 3)				// WS2812's zero high time is long about one third of the period
 #define WS2812_1							(TIMER_PERIOD * 2 / 3)		// WS2812's one high time is long about two thirds of the period
 #define WS2812_RESET						(0)
